Restructure CPU parsing helpers
